swarm-orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarm-orchestrator"
version = "0.1.0"
description = "Conversational Swarm Intelligence Orchestrator for N.O.M.A.D."
requires-python = ">=3.12"
dependencies = [
    "matrix-nio[e2e]>=0.20.0",
    "PyNaCl>=1.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "apscheduler>=3.10.0",
    "tomli>=2.0.0",
    "anyio>=4.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = true
